+++ src/containers/Network/index.js
@@ -3,7 +3,7 @@ import { withTranslation } from 'react-i18next'
 import PropTypes from 'prop-types'
 import { Validators } from './Validators'
 import Nodes from './Nodes'
-import { UpgradeStatus } from './upgradeStatus'
+import { UpgradeStatus } from './UpgradeStatus'
 import { analytics, ANALYTIC_TYPES } from '../shared/utils'
 import './css/style.scss'
 import NoMatch from '../NoMatch'
