+++ src/containers/Network/upgradeStatus.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import axios from 'axios'
 import BarChartVersion from './BarChartVersion'
-import NetworkTabs from './NetworkTab'
+import NetworkTabs from './NetworkTabs'
 import Streams from '../shared/components/Streams'
 import Hexagons from './Hexagons'
 import { localizeNumber } from '../shared/utils'
